+++ Firmament/src/utopia/firmament/image/ButtonImageSet.scala
@@ -4,6 +4,7 @@ import utopia.firmament.model.enumeration.GuiElementState.{Activated, Focused, H
 import utopia.firmament.model.GuiElementStatus
 import utopia.flow.operator.sign.Sign.Negative
 import utopia.genesis.image.Image
+import utopia.paradigm.transform.Adjustment
 
 object ButtonImageSet
 {
@@ -20,23 +21,20 @@ object ButtonImageSet
 	  * @param intensity Brightening intensity modifier (default = 1)
 	  * @return A set that will a) Use 55% alpha while disabled and b) brighten the image on focus and pressed -states
 	  */
-	def brightening(image: Image, intensity: Double = 1) =
-	{
+	def brightening(image: Image, intensity: Double = 1) = {
 		val disabled = image.timesAlpha(0.55)
 		val focused = image.mapEachPixel { _.lightenedBy(intensity) }
 		val pressed = image.mapEachPixel { _.lightenedBy(intensity * 2) }
 		
 		ButtonImageSet(image, focused, pressed, disabled)
 	}
-	
 	/**
 	  * Creates a new button mage set that presents a modified version of the specified image
 	  * @param image The original image
 	  * @param intensity Darkening intensity modifier (default = 1)
 	  * @return A set that will a) Use 55% alpha while disabled and b) darken the image on focus and pressed -states
 	  */
-	def darkening(image: Image, intensity: Double = 1) =
-	{
+	def darkening(image: Image, intensity: Double = 1) = {
 		val disabled = image.timesAlpha(0.55)
 		val focused = image.mapEachPixel { _.darkenedBy(intensity) }
 		val pressed = image.mapEachPixel { _.darkenedBy(intensity) }
@@ -51,8 +49,7 @@ object ButtonImageSet
 	  * @param maxAlpha Alpha used in button pressed state
 	  * @return A button image set that uses varying alpha value
 	  */
-	def varyingAlpha(image: Image, defaultAlpha: Double, maxAlpha: Double) =
-	{
+	def varyingAlpha(image: Image, defaultAlpha: Double, maxAlpha: Double) = {
 		val default = image.timesAlpha(defaultAlpha)
 		val disabled = image.timesAlpha(defaultAlpha * 0.55)
 		val pressed = image.timesAlpha(maxAlpha)
@@ -69,6 +66,25 @@ object ButtonImageSet
 	 */
 	def lowAlphaOnDisabled(image: Image, alphaOnDisabled: Double = 0.55) =
 		ButtonImageSet(image, image, image, image.timesAlpha(alphaOnDisabled))
+	
+	/**
+	  * Creates a button image set where the image changes size based on the button state
+	  * @param image Image for the default state
+	  * @param alphaOnDisabled Alpha modifier for the disabled state (default = 0.55 = 55%)
+	  * @param adjustment Implicit adjustment that determines the size of the size changes
+	  * @return A set that uses the specified image in different sizes
+	  */
+	def changingSize(image: Image, alphaOnDisabled: Double = 0.55)(implicit adjustment: Adjustment) = {
+		// Calculates the maximum image size, which determines the canvas size for all images
+		// Default impact is -1. Focus impact is 0. Activated impact is 0.5 (total).
+		val maxScaling = adjustment(0.5)
+		val newSize = (image.size * maxScaling).round
+		
+		val default = (image * adjustment(-1)).withCanvasSize(newSize)
+		val focus = image.withCanvasSize(newSize)
+		val activated = image.withSize(newSize)
+		apply(default, focus, activated, image.timesAlpha(alphaOnDisabled))
+	}
 }
 
 /**